Refactor: move secrets app cli to separate module
